// src/types/Car.ts
export default interface Car {
    acceleration: number;
    breaking: number;
    carClass: number;
    createdAt: string;
    description: string;
    drive: string;
    engine: string;
    enginePosition: string;
    handling: number;
    images: string[];
    launch: number;
    name: string;
    power: string;
    price: number;
    slug: string;
    speed: number;
    thumbnail: string;
    transmission: string;
    updatedAt: string;
    weight: string;
    _id: string;
}